rename: pr-> pg repository 이름 변경
